requirements.txt
livekit-agents~=1.4
livekit-plugins-deepgram~=1.4
